requirements.txt
Django==3.2.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0